# swagger/swagger.yaml
swagger: '2.0'
info:
  version: 5.5.0
  title: TNTFL
  description: Table Football Ladder
host: foos.cfl.io
schemes:
- http
- https
consumes:
- application/json
produces:
- application/json
paths:
...
  /game/{gameId}/json:
    get:
      parameters:
      - name: gameId
        in: path
        type: number
        format: int64
        description: Timestamp of the game
        required: true
      operationId: getGame
      summary: Get a game
      description: Get a game.
      tags: [games]
      responses:
        200:
          description: Game resource
          schema:
            $ref: '#/definitions/Game'
  /game/add/json:
    post:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - name: redPlayer
        in: query
        type: string
        description: 'The player on red'
        required: true
      - name: redScore
        in: query
        type: integer
        format: int32
        description: 'The red team score'
        required: true
      - name: bluePlayer
        in: query
        type: string
        description: 'The player on blue'
        required: true
      - name: blueScore
        in: query
        type: integer
        format: int32
        description: 'The blue team score'
        required: true
      operationId: addGame
      summary: Add a game
      description: Add a game.
      tags: [games]
      responses:
        200:
          description: A game resource representing the added game
          schema:
            $ref: '#/definitions/Game'
  /games/{begin}/{end}/json:
    get:
      parameters:
      - name: begin
        description: Timestamp to filter from.
        in: path
        type: number
        format: int64
        required: true
      - name: end
        description: Timestamp to filter to.
        in: path
        type: number
        format: int64
        required: true
      - name: includeDeleted
        description: Use value '1' to include deleted games.
        in: query
        type: number
        format: int
        required: false
      operationId: getGames
      summary: Get games
      description: Get games.
      tags: [games]
      responses:
        200:
          description: Games played between given times
          schema:
            type: array
            items:
              $ref: '#/definitions/Game'
  /ladder/json:
    get:
      parameters:
      - name: showInactive
        in: query
        description: Include inactive players
        type: integer
      - name: players
        in: query
        description: Include detailed player info
        type: integer
      operationId: getLadder
      summary: Get the ladder
      description: Get the ladder.
      tags: [ladder]
      responses:
        200:
          description: Ladder
          schema:
            type: array
            items:
              $ref: '#/definitions/LadderEntry'
...
  /speculate/json:
    get:
      parameters:
      - name: showInactive
        in: query
        description: Include inactive players
        type: integer
      - name: players
        in: query
        description: Include detailed player info
        type: integer
      - name: previousGames
        in: query
        description: CSV of speculative games
        type: string
      operationId: speculate
      summary: Get the ladder
      description: Get the ladder.
      tags: [ladder]
      responses:
        200:
          description: Ladder
          schema:
            $ref: '#/definitions/Speculated'
...
definitions:
...
  PlayerTotals:
    type: object
    properties:
      for:
        description: The total number of goals the player has scored
        type: integer
      against:
        description: The total number of goals scored against the player
        type: integer
      wins:
        description: The total number of games the player has won
        type: integer
      losses:
        description: The total number of games the player has lost
        type: integer
      games:
        description: The total number of games the player has played
        type: integer
      gamesToday:
        description: The total number of games the player has played so far today
        type: integer
      gamesAsRed:
        description: The total number of games the player has played as red.
        type: integer
    required: [for, against, wins, losses, games]
  GamesLink:
    type: object
    properties:
      href:
        type: string
    required: [href]
  Player:
    description: Resource representing information about a player
    type: object
    properties:
      name:
        description: Player's cfl username
        type: string
      rank:
        description: Player's current ladder rank
        type: integer
        format: int32
      games:
        description: Link to get all games for this player
        $ref: '#/definitions/GamesLink'
      activity:
        description: Player's activity level
        type: number
      skill:
        description: The player's skill value in the ranking system
        type: number
        format: double
      total:
        description: Statistics for the player
        $ref: '#/definitions/PlayerTotals'
    required: [name, rank, games, activity, skill, total]
  Performance:
    description: Resource representing how a player performed in a game
    type: object
    properties:
      name:
        description: Player's cfl username
        type: string
      score:
        description: Goals scored by this player
        type: integer
        format: int32
      skillChange:
        description: Skill value change for this player
        type: number
        format: double
      rankChange:
        description: Ladder rank change for this player
        type: integer
        format: int32
      newRank:
        description: New ladder rank of this player
        type: integer
        format: int32
      href:
        description: Link to get player's information
        type: string
      achievements:
        description: Achievements earned during this performance
        type: array
        items:
          $ref: '#/definitions/Achievement'
    required: [name, score, skillChange, rankChange, newRank, href, achievements]
  Achievement:
    type: object
    properties:
      name:
        description: Name of the achievement
        type: string
      description:
        description: A brief description of the achievement
        type: string
      time:
        description: Timestamp of when the achievement was earned
        type: integer
        format: int64
    required: [name, description]
  Game:
    type: object
    properties:
      date:
        description: The unix timestamp of when the game was added
        type: integer
        format: int64
      positionSwap:
        description: Whether or not the players swapped ladder position
        type: boolean
      blue:
        $ref: '#/definitions/Performance'
      red:
        $ref: '#/definitions/Performance'
      deleted:
        type: object
        properties:
          at:
            description: Unix time game was deleted at
            type: integer
            format: int64
          by:
            description: User that deleted the game
            type: string
        required: [at, by]
    required: [date, positionSwap, blue, red]
  TrendItem:
    type: object
    properties:
      date:
        description: Unix time
        type: number
        format: int64
      skill:
        type: number
        format: float
    required: [date, skill]
  LadderEntry:
    type: object
    properties:
      rank:
        description: Player's rank
        type: integer
      name:
        description: Player's name
        type: string
      skill:
        description: Player's skill
        type: number
      href:
        description: Link to detailed player stats
        type: string
      player:
        $ref: '#/definitions/Player'
      trend:
        description: Player's skill point trend
        type: array
        items:
          $ref: '#/definitions/TrendItem'
...
  Speculated:
    type: object
    properties:
      entries:
        description: Ladder entries
        type: array
        items:
          $ref: '#/definitions/LadderEntry'
      games:
        description: Ladder entries
        type: array
        items:
          $ref: '#/definitions/Game'
    required: [entries, games]
...